Added handler for local Git repo and [ecode]
Still working out what the code browser display will look like.
Thinking it would be nice to allow the header and footer format to be
set as a config option.  But want users to be able to easily toggle
hiding, partially showing, and fully showing code boxes, and the admin
to set how many lines are displayed if a code box is partially shown.